+++ src/bundled-web-search-registry.ts
@@ -1,4 +1,5 @@
 import bravePlugin from "../extensions/brave/index.js";
+import exaPlugin from "../extensions/exa/index.js";
 import firecrawlPlugin from "../extensions/firecrawl/index.js";
 import googlePlugin from "../extensions/google/index.js";
 import moonshotPlugin from "../extensions/moonshot/index.js";
@@ -21,6 +22,12 @@ export const bundledWebSearchPluginRegistrations: ReadonlyArray<{
       return bravePlugin;
     },
     credentialValue: "BSA-test",
+  },
+  {
+    get plugin() {
+      return exaPlugin;
+    },
+    credentialValue: "exa-test",
   },
   {
     get plugin() {