+++ src/plugins/bundled-web-search-ids.ts
@@ -1,5 +1,6 @@
 export const BUNDLED_WEB_SEARCH_PLUGIN_IDS = [
   "brave",
+  "exa",
   "firecrawl",
   "google",
   "moonshot",